Enable ty type checking on test directories
- Add packages/*/tests to [tool.ty.src] include paths
- Fix poll_interval/timeout types: int -> float in SlurmJobManager
- Fix TrackingReader.__iter__ return type: Iterable -> Iterator
- Wrap string args with Path() in RunState.create test calls
- Add non-None assertions before  checks on optional types
- Switch type ignore comments to ty syntax (# ty: ignore[...])

+++ packages/jernerics/src/jernerics/hpc/slurm.py
@@ -170,7 +170,7 @@ class SlurmJobManager:
         return None
 
     def wait_for_completion(
-        self, job_id: str, poll_interval: int = 30, timeout: int | None = None
+        self, job_id: str, poll_interval: float = 30, timeout: float | None = None
     ) -> bool:
         import time
